refactor: move city lights provider to services
