+++ packages/photo-companion/src/routes/now/components/now-lights.tsx
@@ -1,7 +1,7 @@
 import { CardEntry, CardInfo, SupportsLights } from "~/lib/components";
 import { LightsCountdown } from "~/lib/components/lights-countdown/lights-countdown";
-import { useCityLights } from "~/lib/context/city-lights";
 import { useTranslation } from "~/lib/context/translation";
+import { useCityLights } from "~/services/city-lights";
 
 export const NowLightsInfo = () => {
 	const { formatters, t } = useTranslation();
