+++ packages/photo-companion/src/routes/timeline/timeline.page.tsx
@@ -5,9 +5,9 @@ import { incrementDateByDay } from "utils/date";
 import type { EventGroupName } from "types";
 
 import { Timeline, TimelineEvent, TimelineEventEmpty, TimelineGroup } from "~/lib/components";
-import { CityLightsProvider } from "~/lib/context/city-lights";
 import { useTranslation } from "~/lib/context/translation";
 import { useDatetime } from "~/lib/hooks";
+import { CityLightsProvider } from "~/services/city-lights";
 import { useSettings } from "~/services/settings";
 
 import { useDisabledTimeline, useTimelineEvents } from "./hooks";
